+++ src/file/ident.rs
@@ -15,6 +15,7 @@ use concat_string::concat_string;
 /// let ident = get_ident("test", "20190101");
 /// assert_eq!(ident, "test20190101");
 /// ```
+#[deprecated(since = "0.4.0", note = "We don't use this identifier anymore.")]
 pub fn get_ident(filename: &str, timestamp: &str) -> String {
     concat_string!(filename, timestamp)
 }